requirements.txt
cachetools==3.0.0
google-api-python-client==1.7.5
google-auth==1.6.1
google-auth-httplib2==0.0.3
httplib2==0.12.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
rsa==4.0
six==1.11.0
uritemplate==3.0.0
Click==7.0
gunicorn==19.7.1
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
Pillow==5.3.0
MarkupSafe==1.1.0
pytesseract==0.2.5
Werkzeug==0.14.1
opencv-python==3.3.0.9
